refactor code: - move device and host in their own module + change their functions prototype, remove the name from function, for example isnteed of pa::get_device_info, use pa::device::get_info -  clean comments format to follow rust conventions
